Event count reads now enabled in SelectionCriteria

+++ cpp/libs/opendnp3/src/opendnp3/outstation/EventResponseContext.h
@@ -47,7 +47,7 @@ public:
 
 	IINField ReadAll(const GroupVariationRecord& record);
 
-	IINField ReadCount(const GroupVariationRecord& record, uint32_t count);
+	IINField ReadCount(const GroupVariationRecord& record, uint16_t count);
 
 	void Reset();
